+++ src/bvh.h
@@ -10,11 +10,11 @@
 class BVHNode : public Hitable {
 public:
 
-    BVHNode(std::vector<std::shared_ptr<Hitable>> list, double t0, double t1);
+    BVHNode(std::vector<std::shared_ptr<Hitable>> list, float t0, float t1);
 
-    bool hit(Ray const& r, double t_min, double t_max,
+    bool hit(Ray const& r, float t_min, float t_max,
              HitRecord& rec) const override;
-    bool bounding_box(double t0, double t1, AABB& bbox) const override;
+    bool bounding_box(float t0, float t1, AABB& bbox) const override;
 
 private:
     std::shared_ptr<Hitable> _left;
