+++ src/sphere.h
@@ -8,15 +8,15 @@ class Material;
 
 class Sphere : public Hitable {
 public:
-    Sphere(Vec3 const& center, double radius, Material* mat_ptr);
+    Sphere(Vec3 const& center, float radius, Material* mat_ptr);
 
-    bool hit(Ray const& r, double t_min, double t_max,
+    bool hit(Ray const& r, float t_min, float t_max,
              HitRecord& rec) const override;
-    bool bounding_box(double t0, double t1, AABB& box) const override;
+    bool bounding_box(float t0, float t1, AABB& box) const override;
 
 private:
     Vec3 center;
-    double radius;
+    float radius;
     Material* mat_ptr;
 };
